Ad views finished. Modified Article model
To allow users to upload pictures of articles a string property
was added to the Article model containing the path of such image.
Solved the error thrown when trying to buy, edit or delete an article;
these links in the Ad/Details view were passing the user's name as the Id,
apparently when authentification was implemented the views were not
updated.