Shopping Cart views finished
AddToShoppingCart view should be deleted, now you can add an article to
the shopping cart directly from that ad's Details view. A button was
placed at the ShoppingCart/Index view to delete the current shopping list
but it's implementation is pending (maybe this button is unnecessary).
There are some implementation issues at the modal form shown when adding
an article to the shopping cart; it works fine but I don't think is
correct.