Restrict ads removal and edition to creators only.
Know and user can edit and or delete an ad only they created it.
Initialy a different page was going to be created in order to display MyAds,
but due to filters and the amount of stuff this would've require to be changed
a different approach was taken into account. A mystuffFilter shoul be ceated if
this approach is the one we continue with.