Complete of my Stuff pages
My Adds And My Auctions pages were created. A new filter for
state of the auction(closed, in curse, pending) was added. Index pages
of Adds and Auctions now show adds with more than 1 article available
and auctions in curse or pending. All filter was modified in order to show
all adds and auctions.